Consolidate index.js to only exports
